+++ Core/Matcher/ContentMatcher.php
@@ -25,8 +25,8 @@ class ContentMatcher extends QueryBasedMatcher implements SortingMatcherInterfac
         self::MATCH_CONTENT_ID, self::MATCH_LOCATION_ID, self::MATCH_CONTENT_REMOTE_ID, self::MATCH_LOCATION_REMOTE_ID,
         self::MATCH_ATTRIBUTE, self::MATCH_CONTENT_TYPE_ID, self::MATCH_CONTENT_TYPE_IDENTIFIER, self::MATCH_GROUP,
         self::MATCH_CREATION_DATE, self::MATCH_MODIFICATION_DATE, self::MATCH_OBJECT_STATE, self::MATCH_OWNER,
-        self::MATCH_PARENT_LOCATION_ID, self::MATCH_PARENT_LOCATION_REMOTE_ID, self::MATCH_SECTION, self::MATCH_SUBTREE,
-        self::MATCH_VISIBILITY,
+        self::MATCH_PARENT_LOCATION_ID, self::MATCH_PARENT_LOCATION_REMOTE_ID, self::MATCH_QUERY_TYPE, self::MATCH_SECTION,
+        self::MATCH_SUBTREE, self::MATCH_VISIBILITY,
         // aliases
         'content_type',
         // BC
@@ -135,11 +135,17 @@ class ContentMatcher extends QueryBasedMatcher implements SortingMatcherInterfac
                         }
                     }
 
-                    $query = new Query();
+                    if ($key == self::MATCH_QUERY_TYPE) {
+                        $query = $this->getQueryByQueryType($values);
+                    } else {
+                        $query = new Query();
+                        $query->filter = $this->getQueryCriterion($key, $values);
+                    }
+
+                    // q: when getting a query via QueryType, should we always inject offset/limit?
                     $query->limit = $limit != 0 ? $limit : $this->queryLimit;
                     $query->offset = $offset;
                     if (isset($query->performCount)) $query->performCount = false;
-                    $query->filter = $this->getQueryCriterion($key, $values);
                     if (!empty($sort)) {
                         $query->sortClauses = $this->getSortClauses($sort);
                     } else {