+++ WHATSNEW.md
@@ -1,6 +1,9 @@
 Version XX (unreleased)
 =======================
 
+* New: Content and Location matchers, used in `load`, `update` and `delete` steps for Content and Location can now match
+  by QueryType (issue #239)
+
 * New: taught the `kaliop:migration:status` command to display full migration path by using the `--show-path` option
   (issue #152)